fixed problem3

+++ public_html/M2/problem3.php
@@ -9,7 +9,7 @@ function bePositive($arr) {
     echo "<br>Positive output:<br>";
     //TODO use echo to output all of the values as positive (even if they were originally positive)
     foreach ( $arr as $i){
-        echo abs(intval($i));
+        echo abs(floatval($i));
     }
 
 }
